src/main/java/com/tardin/bookstoremanager/users/service/AuthenticationService.java
package com.tardin.bookstoremanager.users.service;

import com.tardin.bookstoremanager.users.dto.AuthenticatedUser;
import com.tardin.bookstoremanager.users.entity.User;
import com.tardin.bookstoremanager.users.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import static java.lang.String.format;

@Service
public class AuthenticationService implements UserDetailsService {

    @Autowired
    private UserRepository repository;

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        User user =  repository.findByUsername(username)
                .orElseThrow(() -> new UsernameNotFoundException(format("User not found with username %s", username)));
        return new AuthenticatedUser(user.getUsername(), user.getPassword(), user.getRole().getDescription());
    }
}
